treino carregando completamente

+++ src/principal/Aerobico.java
@@ -1,23 +1,23 @@
 package principal;
 
 public class Aerobico extends Exercicio{
-    private float tempoMinutos;
+    private int tempoMinutos;
 
     public Aerobico(){
         super();
         tempoMinutos = 0;
     }
 
-    public Aerobico(String nome, float tempoMinutos) {
+    public Aerobico(String nome, int tempoMinutos) {
         super(nome);
         this.tempoMinutos = tempoMinutos;
     }
 
-    public float getTempoMinutos() {
+    public int getTempoMinutos() {
         return tempoMinutos;
     }
 
-    public void setTempoMinutos(float tempoMinutos) {
+    public void setTempoMinutos(int tempoMinutos) {
         this.tempoMinutos = tempoMinutos;
     }
 }
